+++ tests/testthat/test-51-pxmake-valid-px-file.R
@@ -104,5 +104,37 @@ test_that("px file is valid", {
     expect_that_px_filehas_timeval(fotest)
   })
 
+  test_that("axis-version is quoted", {
+    expect_that_axis_version_is_quoted <- function(path) {
+      px_lines <- readLines(path)
+
+      n_quotes <-
+        stringr::str_count(px_lines[stringr::str_detect(px_lines, "^AXIS-VERSION")],
+                  '"'
+                  )
+
+      expect_equal(n_quotes, 2)
+    }
+
+    expect_that_axis_version_is_quoted(bexsta)
+    expect_that_axis_version_is_quoted(bexltall)
+    expect_that_axis_version_is_quoted(fotest)
+  })
+
+  test_that("VARIABLECODE has variable", {
+    expect_that_variablecode_has_varaible <- function(path) {
+      px_lines <- readLines(path)
+
+      n_parentheses <-
+        stringr::str_count(px_lines[stringr::str_detect(px_lines, "^VARIABLECODE")],
+                           "\\("
+                           )
+
+      expect_true(all(n_parentheses > 0))
+    }
+
+    expect_that_variablecode_has_varaible(bexsta)
+  })
+
   expect_true(TRUE) #needed to run
 })